+++ config/imager-x-cloudflare-images-transformer.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    'zoneDomain' => 'media.studiosyntax.nl',
+    'defaultParams' => [
+        'format' => 'auto',
+        'quality' => 80
+    ],
+    'autoGravityWhenNoFocalPoint' => true
+];